refactor: moved ToJson to internal place so is not exposed to user
